fix(scheduler): truncate timesteps to integer matching Python int64 dtype (#129)
Python converts timesteps via torch.int64 which truncates 999.8 to 999.
Swift was keeping the full float. The 0.8 difference causes up to 72.5%
error in high-frequency sinusoidal position embeddings, biasing every
transformer block's AdaLN modulation. UniPC's multi-step D1 history then
amplifies the per-step bias into exponential latent growth (divergence).

+++ Sources/ZImage/Wan/Pipeline/WanI2VPipeline.swift
@@ -272,8 +272,12 @@ public final class WanI2VPipeline {
       let model = try moeManager.model(forTimestep: t)
       let guideScale = moeManager.guideScale(forTimestep: t, scales: config.guideScale)
 
-      // Prepare timestep tensor
-      let timestep = MLXArray([t])
+      // Prepare timestep tensor — truncate to integer to match Python's int64 dtype.
+      // The sinusoidal position embedding is sensitive to fractional differences;
+      // passing 999.8 instead of 999 shifts high-frequency components enough to
+      // bias the noise prediction, which the multi-step UniPC solver then amplifies
+      // into divergent latent growth.
+      let timestep = MLXArray([Float(Int32(t))])
 
       // CFG: conditional forward pass
       let noisePredCond = model.forward(
